+++ prophet/plugins/http/example/HttpExample.sol
@@ -3,11 +3,13 @@ pragma solidity ^0.8.25;
 
 import "precompile-json/IJson.sol" as json;
 import "precompile-async/IAsync.sol" as async;
+import "precompile-sched/ISched.sol" as sched;
+import "precompile-callbacks/ICallback.sol" as callback;
 import "precompile-common/Types.sol" as types;
 import "http/Http.sol" as http;
 
-contract HttpExample {
-    uint64 public lastFutureId;
+contract HttpExample is callback.ICallback {
+    uint64 public lastTaskId;
 
     int256 public bitcoinPrice;
     int256 public tetherPrice;
@@ -15,20 +17,32 @@ contract HttpExample {
 
     bytes public responseBody;
 
+    // Receive function to accept award transfers
+    receive() external payable {
+        // Contract can now receive award
+        // You can add custom logic here if needed
+    }
+
     function run() public returns (http.Http.Request memory request) {
         request.url = "https://api.coingecko.com/api/v3/simple/price?ids=bitcoin,tether,uniswap&vs_currencies=usd";
         request.method = "GET";
         request.body = "";
 
-        lastFutureId = async.IASYNC_CONTRACT.addTask("http", abi.encode(request), address(this));
+        types.Types.Coin[] memory maxFees = new types.Types.Coin[](1);
+        maxFees[0] = types.Types.Coin("award", 1);
+
+        uint64 maxGas = 1000000000;
+        sched.CallbackParams memory callbackData = sched.CallbackParams(address(this), maxGas);
+
+        lastTaskId = async.IASYNC_CONTRACT.addTask("http", abi.encode(request), maxFees, callbackData);
 
         bitcoinPrice = 0;
         tetherPrice = 0;
         uniswapPrice = 0;
     }
 
-    function cb() external {
-        async.TaskByIdResponse memory future = async.IASYNC_CONTRACT.taskById(lastFutureId);
+    function cb(uint64 id, bytes calldata output) external override returns (bytes memory) {
+        async.TaskByIdResponse memory future = async.IASYNC_CONTRACT.taskById(lastTaskId);
         if (future.taskResponse.result.id == 0) {
             revert("Not ready yet");
         }
@@ -56,4 +70,7 @@ contract HttpExample {
         tetherPrice = tether;
         uniswapPrice = uniswap;
     }
+
+
+
 }