+++ tests/cases/act_precompile.go
@@ -26,14 +26,11 @@ type Test_ActPrecompile struct {
 	w *exec.WardenNode
 }
 
-func (c *Test_ActPrecompile) Setup(t *testing.T, build framework.BuildResult) {
-	c.w = exec.NewWardenNode(t, build.Wardend)
-
-	go c.w.Start(t, "./testdata/snapshot-many-users")
-	c.w.WaitRunning(t)
+func (c *Test_ActPrecompile) Setup(t *testing.T, f *framework.F) {
+	c.w = f.StartNodeFromSnapshot(t, framework.SnapshotPrecompiles)
 }
 
-func (c *Test_ActPrecompile) Run(t *testing.T, _ framework.BuildResult) {
+func (c *Test_ActPrecompile) Run(t *testing.T, _ *framework.F) {
 	alice := exec.NewWardend(c.w, "alice")
 	bob := exec.NewWardend(c.w, "bob")
 	dave := exec.NewWardend(c.w, "dave")