+++ tests/cases/json_precompile.go
@@ -23,14 +23,11 @@ type Test_JsonPrecompile struct {
 	w *exec.WardenNode
 }
 
-func (c *Test_JsonPrecompile) Setup(t *testing.T, build framework.BuildResult) {
-	c.w = exec.NewWardenNode(t, build.Wardend)
-
-	go c.w.Start(t, "./testdata/snapshot-many-users")
-	c.w.WaitRunning(t)
+func (c *Test_JsonPrecompile) Setup(t *testing.T, f *framework.F) {
+	c.w = f.StartNodeFromSnapshot(t, framework.SnapshotPrecompiles)
 }
 
-func (c *Test_JsonPrecompile) Run(t *testing.T, _ framework.BuildResult) {
+func (c *Test_JsonPrecompile) Run(t *testing.T, _ *framework.F) {
 	alice := exec.NewWardend(c.w, "alice")
 
 	evmClient := c.w.EthClient(t)