+++ tests/cases/keychain_writers.go
@@ -23,14 +23,11 @@ type Test_KeychainWriters struct {
 	w *exec.WardenNode
 }
 
-func (c *Test_KeychainWriters) Setup(t *testing.T, build framework.BuildResult) {
-	c.w = exec.NewWardenNode(t, build.Wardend)
-
-	go c.w.Start(t, "./testdata/snapshot-keychain")
-	c.w.WaitRunning(t)
+func (c *Test_KeychainWriters) Setup(t *testing.T, f *framework.F) {
+	c.w = f.StartNodeFromSnapshot(t, framework.SnapshotKeychain)
 }
 
-func (c *Test_KeychainWriters) Run(t *testing.T, build framework.BuildResult) {
+func (c *Test_KeychainWriters) Run(t *testing.T, f *framework.F) {
 	bob := exec.NewWardend(c.w, "bob")
 	writer := exec.NewWardend(c.w, "writer")
 	client := TestGRPCClient(*c.w.GRPCClient(t))