+++ tests/cases/owner_approve_action.go
@@ -20,14 +20,11 @@ type Test_OwnerApproveAction struct {
 	w *exec.WardenNode
 }
 
-func (c *Test_OwnerApproveAction) Setup(t *testing.T, build framework.BuildResult) {
-	c.w = exec.NewWardenNode(t, build.Wardend)
-
-	go c.w.Start(t, "./testdata/snapshot-many-users")
-	c.w.WaitRunning(t)
+func (c *Test_OwnerApproveAction) Setup(t *testing.T, f *framework.F) {
+	c.w = f.StartNodeFromSnapshot(t, framework.SnapshotManyUsers)
 }
 
-func (c *Test_OwnerApproveAction) Run(t *testing.T, _ framework.BuildResult) {
+func (c *Test_OwnerApproveAction) Run(t *testing.T, _ *framework.F) {
 	client := TestGRPCClient(*c.w.GRPCClient(t))
 
 	alice := exec.NewWardend(c.w, "alice")