+++ tests/cases/warden_precompile.go
@@ -24,16 +24,12 @@ type Test_WardenPrecompile struct {
 	w *exec.WardenNode
 }
 
-func (c *Test_WardenPrecompile) Setup(t *testing.T, build framework.BuildResult) {
-	c.w = exec.NewWardenNode(t, build.Wardend)
-
-	go c.w.Start(t, "./testdata/snapshot-many-users")
-	c.w.WaitRunning(t)
-
+func (c *Test_WardenPrecompile) Setup(t *testing.T, f *framework.F) {
+	c.w = f.StartNodeFromSnapshot(t, framework.SnapshotPrecompiles)
 	c.w.PrintLogsAtTheEnd(t, []string{"TEST_DEBUG"})
 }
 
-func (c *Test_WardenPrecompile) Run(t *testing.T, build framework.BuildResult) {
+func (c *Test_WardenPrecompile) Run(t *testing.T, f *framework.F) {
 	alice := exec.NewWardend(c.w, "alice")
 	bob := exec.NewWardend(c.w, "bob")
 	dave := exec.NewWardend(c.w, "dave")