+++ CHANGELOG.md
@@ -62,6 +62,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Bug Fixes
 - (x/async) PFP image generation plugin will only take one input (string) which is the AI prompt
 
+- (x/async) Fix missing Solver field in precompile queries
+
 ## [v0.6.3](https://github.com/warden-protocol/wardenprotocol/releases/tag/v0.6.3) - 2025-04-16
 
 ### Features (non-breaking)